poc_extras/functions_pubsub_serveless/custom_role_add_member_function/requirements.txt
google-api-python-client==2.11.0
google-auth==1.31.0
google-auth-httplib2==0.1.0

